\name{qtQDA-package}
\docType{package}
\alias{qtQDA-package}
\title{qtQDA: Quantile Transformation and Gaussian Quadratic Discriminant Analysis}

\description{
This package contains functions for classification, quantile transformation and quadratic discrimant analysis and several generalized linear model analysis of RNA-Seq data. 

For complete list of functions, use library(help = "qtQDA").
}

\author{
Necla Kayaalp, Goknur Giner, Gordon Smyth and Luke Gandolfo
}

\keyword{package}
